move validation/ to shared/
