+++ shared/validation/__test__/Coordinates.test.ts
