+++ shared/validation/inputValidation/ValidateCoordinates.ts
