+++ shared/validation/inputValidation/ValidateHashtag.ts
