+++ shared/validation/inputValidation/ValidateVerificationToken.ts
