+++ src/mutations/VerifyToken.ts
@@ -1,4 +1,4 @@
-import { ValidateVerificationToken } from '../validation';
+import { ValidateVerificationToken } from '../../shared/validation';
 import RunCypherQuery from '../database/RunCypherQuery';
 
 const VerifyUser = async (token: string): Promise<string> => {
