+++ tsconfig-express.json
@@ -1,10 +1,5 @@
 {
   "compilerOptions": {
-    "baseUrl": "./",
-    "paths": {
-      "@src/*": ["src/*"],
-      "@validation/*": ["src/validation/*"]
-    },
     "module": "commonjs",
     "outDir": "dist",
     "sourceMap": true,
@@ -14,6 +9,6 @@
     "skipLibCheck": true
   },
   "compileOnSave": true,
-  "include": ["src"],
+  "include": ["src", "shared"],
   "exclude": ["**/__test__"]
 }
